test: lift tenant placement util
